pyproject.toml
[project]
name = "literature-review-agent"
version = "0.1.0"
description = "LangGraph-powered multi-agent literature review system"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "LitRev AI" }
]

dependencies = [
  "langchain>=0.2.10",
  "langgraph>=0.2.0",
  "pydantic>=2.6",
  "tenacity>=8.2",
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "python-dotenv>=1.0",
  "typer>=0.12",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "jinja2>=3.1",
  "rich>=13.7",
  "tqdm>=4.66",
  "langchain-openai>=0.1.7; python_version >= '3.11'",
  "langchain-anthropic>=0.1.15; python_version >= '3.11'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-mock>=3.14",
  "pytest-asyncio>=0.23",
  "ruff>=0.6.8",
  "black>=24.8",
  "mypy>=1.11",
  "types-requests",
]

[project.scripts]
litrev = "src.cli.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "TID", "SIM", "PLE", "PLW"]
ignore = ["E203", "E266", "E501"]
src = ["src", "cli", "api", "tests"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
strict_optional = true
ignore_missing_imports = true
namespace_packages = true
explicit_package_bases = true
mypy_path = ["."]


